requirements.txt
typer[all]
httpx~=0.28.1
openai
pdfminer.six
rich
PyPDF2~=3.0.1
python-docx~=1.2.0
python-dotenv~=1.1.1